Refactor MemberList into managers
