+++ logs/log1756064359.md
@@ -1,0 +1,5 @@
+- Refactored MemberList by introducing dedicated managers:
+  - Added ConsensusManager to encapsulate topology consensus operations.
+  - Added MemberStrategyManager to maintain per-kind member strategies and activator lookups.
+  - Simplified MemberList to delegate consensus and strategy responsibilities, reducing coupling and clarifying responsibilities.
+- All core tests (Proto.Actor, Proto.Remote, Proto.Cluster) executed successfully after the refactor.
